[FLINK-24060][Tests]Move ZooKeeperUtilTest to right class
